test generator, configs and nonfunctional refactoring

+++ src/main/scala/warehouse/Warehouse.scala
@@ -5,28 +5,31 @@ import org.apache.spark.sql.types._
 import org.apache.spark.sql.functions._
 import warehouse.models.{Amount, Position}
 import warehouse.spark.LastValueByTimestamp
+import warehouse.utils.Configured
 
 import scala.reflect.io.Path
 
-object Warehouse {
-  val POSITIONS_PATH = "/positions.csv"
-  val AMOUNTS_PATH = "/amounts.csv"
+object Warehouse extends Configured {
+  val PositionsPath: String = config.getString("warehouse.positions_path")
+  val AmountsPath: String = config.getString("warehouse.amounts_path")
+  val SaveCurrentAmountsPath: String = config.getString("warehouse.save_current_amounts_path")
+  val SaveMaxMinAvgPath: String = config.getString("warehouse.save_max_min_avg_path")
 
-  val SAVE_CURRENT_AMOUNTS_PATH = "current_amounts.csv"
-  val SAVE_MAX_MIN_AVG_PATH = "max_min_avg.csv"
-
-  Seq(SAVE_CURRENT_AMOUNTS_PATH, SAVE_MAX_MIN_AVG_PATH).foreach { path =>
-    Path(path).deleteRecursively()
+  if (config.getBoolean("warehouse.delete_prev_results")) {
+    Seq(SaveCurrentAmountsPath, SaveMaxMinAvgPath).foreach { path =>
+      Path(path).deleteRecursively()
+    }
   }
 
-  val positionsPath: String = locateResource(POSITIONS_PATH)
-  val amountsPath: String = locateResource(AMOUNTS_PATH)
+  val locatedPositionsPath: String = locateResource(PositionsPath)
+  val locatedAmountsPath: String = locateResource(AmountsPath)
 
   val spark: SparkSession =
     SparkSession
       .builder()
       .appName("Warehouse Statistics")
-      .config("spark.master", "local")
+      .config("spark.master", "local[4]")
+      .config("spark.speculation", "true")
       .getOrCreate()
 
   import spark.implicits._
@@ -34,8 +37,8 @@ object Warehouse {
   def main(args: Array[String]): Unit = {
     try {
       // Load required data from files, e.g. csv of json.
-      val positionsDS: Dataset[Position] = readCsv(positionsPath, Position.schema, Position.apply)
-      val amountsDS: Dataset[Amount] = readCsv(amountsPath, Amount.schema, Amount.apply)
+      val positionsDS: Dataset[Position] = readCsv(locatedPositionsPath, Position.schema, Position.apply)
+      val amountsDS: Dataset[Amount] = readCsv(locatedAmountsPath, Amount.schema, Amount.apply)
 
       //region Find the current amount for each position, warehouse, product.
 
@@ -50,8 +53,6 @@ object Warehouse {
           (position.positionId, position.warehouse, position.product, currentAmount)
         }
       //endregion
-
-      positionWithCurrentAmount.show()
 
       //region Find max, min, avg amounts for each warehouse and product
 
@@ -74,18 +75,29 @@ object Warehouse {
         }
       //endregion
 
-      result2.show()
-
       // saving
-      result1.write.csv(SAVE_CURRENT_AMOUNTS_PATH)
-      result2.write.csv(SAVE_MAX_MIN_AVG_PATH)
+      result1.write.csv(SaveCurrentAmountsPath)
+      result2.write.csv(SaveMaxMinAvgPath)
 
+//      // showing
+//      result1.show()
+//      result2.show()
+//
+//      // detailed
+//      println(s"Result1: ${result1.rdd.toDebugString}")
+//      println(s"Result2: ${result2.rdd.toDebugString}")
+
+    } catch {
+      case any: Throwable =>
+        any.printStackTrace()
     } finally {
+      println("Press ENTER to close the app")
+      scala.io.StdIn.readLine()
       spark.stop()
     }
   }
 
-  def locateResource(path: String): String = getClass.getResource(path).getPath
+  def locateResource(path: String): String = getClass.getClassLoader.getResource(path).getPath
 
   def readCsv[T : Encoder](path: String, schema: StructType, toTypedFoo: Row => T): Dataset[T] = {
     spark.read
@@ -96,7 +108,6 @@ object Warehouse {
       .option("timestampFormat", "s") // epoch time in seconds
       .csv(path)
       .map(toTypedFoo)
-      .cache()
   }
 
   def currentAmounts(amounts: Dataset[Amount]): Dataset[(Long, java.math.BigDecimal)] = {
